backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pypdf
sentence-transformers
numpy
langchain-text-splitters
google-genai
python-dotenv
pydantic
faiss-cpu